Fix issues related to scalar 32-bit MSVC builds.
+ In `CMakeAddFortranSubdirectory.cmake`, only munge MPI settings if
  `DRACO_C4=MPI`.
+ In `windows-cl.cmake`:
  - Use `CMAKE_SIZEOF_VOID_P` instead of `CMAKE_CL_64` as recommended by CMake.
  - Make searching for `ws2_32.lib` more robust by searching in more locations
    and exiting the seearch loop once found.
+ In `FortranChecks/test/CMakeLists.txt`, make the search for
  `libquadmath-0.dll` more robust.
+ Fix some comparisons between signed and unsigned integral values in
  `RTT_Format_Reader` and in `mdpsan`'s implementation.
+ Suppress a warning about unreachable code in `C4_Serial.cc`.

+++ src/RTT_Format_Reader/Flags.hh
@@ -50,11 +50,11 @@ public:
   /*!
    * \brief Validates the specified flag index.
    * \param flag Flag index.
-   * \return The existance of the flag.
+   * \return The existence of the flag.
    */
   bool allowed_flag(size_t flag) const {
-    return flag_nums.end() !=
-           std::find(flag_nums.begin(), flag_nums.end(), flag);
+    return flag_nums.end() != std::find(flag_nums.begin(), flag_nums.end(),
+                                        static_cast<int>(flag));
   }
 
   /*!
